Adds fixes to run page specific js (#183)
`admin_app`
-----------
- Make necessary changes to make page specific js
