+++ apps/admin_app/assets/js/views/main.js
@@ -1,0 +1,11 @@
+export default class MainView {
+    mount() {
+      // This will be executed when the document loads...
+      console.log('MainView mounted');
+    }
+  
+    unmount() {
+      // This will be executed when the document unloads...
+      console.log('MainView unmounted');
+    }
+}
